fix: fix the camera title and decoupling the code
1. add the camera title
2. decoupling the code
